# Conflicts:
#	maplibre/src/headless.rs
#	maplibre/src/render/stages/upload_stage.rs
#	maplibre/src/stages/request_stage.rs

+++ maplibre/src/render/tile_view_pattern.rs
@@ -41,7 +41,7 @@ impl TileShape {
             coords,
             zoom_factor: zoom.scale_to_tile(&coords),
             transform: coords.transform_for_zoom(zoom),
-            buffer_range: index as u64 * STRIDE..(index as u64 + 1) * STRIDE,
+            buffer_range: index * STRIDE..(index + 1) * STRIDE,
         }
     }
 }
